Make code more compact
